Add car/car_logger
Add a client tool for collecting data from a vehicle's OBD-II interface
and publishing it to an MQTT broker. The tool is to be run on a computer
connected to the vehicle OBD-II interface.

The code contains filtering applied on-the-spot during the final test
drives, which can be removed without breaking the functionality.