Add mqtt-logger.k8s.yaml
Add an OKD Kubernetes yaml configuration file for deploying the MQTT
logger, the InfluxDB timeseries database, and a Grafana instance. All of
these are to be deployed with the necessary configuration files,
secrets, persistent volume claims (PVC), services and routes,

The deployment of the MQTT logger is dependent on having its image
available in some image registry and referenced properly from the K8S
config file. In addition, all of the secrets have been removed from the
config for security reasons, which does break the functionality of the
config.